Prevent annoying signed/unsigned comparison warning
Commit d6c013f303 introduced a cast to avoid a signed/unsigned
comparison warning for all translation units that included
core/dive.h.

Commit 1f8506ce64 then changed the definition of duration_t from
unsigned to signed, inverting the effect of d6c013f303.

Thus, revert d6c013f303 to allow compilation with -Wall without
flooding.

+++ core/dive.h
@@ -963,7 +963,7 @@ static inline struct gasmix *get_gasmix(struct dive *dive, struct divecomputer *
 		gasmix = &dive->cylinder[cyl].gasmix;
 		ev = dc ? dc->events : NULL;
 	}
-	while (ev && ev->time.seconds < (unsigned int)time) {
+	while (ev && ev->time.seconds < time) {
 		gasmix = get_gasmix_from_event(dive, ev);
 		ev = get_next_event(ev->next, "gaschange");
 	}
